+++ lib/mx_request.dart
@@ -3,9 +3,8 @@
 /// More dartdocs go here.
 library;
 
-export 'package:dio/dio.dart';
 export 'src/request_content/request_content.dart';
 export 'src/annotation.dart';
 export 'src/request_base.dart';
 export 'src/request_method.dart';
-export 'src/dio_extension.dart';
+export 'src/extension/extension.dart';